pyproject.toml
[project]
name = "clausebot_api"
version = "0.1.0"
description = "AWS D1.1 Welding Inspection and Quiz Generation API"
requires-python = ">=3.10"
authors = [
    {name = "MiltmonNDT", email = "you@example.com"}
]
readme = "README.md"

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "supabase>=2.6.0",
    "pydantic>=2.8.2",
    "requests>=2.31.0",
    "pyairtable>=2.3.3",
    "aiohttp>=3.9.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["clausebot_api*"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3
